WIP
Summary: minor improvements

Test Plan: manual

Reviewers: feng

Reviewed By: feng

Differential Revision: https://phab.clearcutcorp.com/D25

+++ parser/run_parallel.sh
@@ -10,21 +10,49 @@ if [ "$#" -le 1 ]; then
   exit
 fi
 
-INPUT_FILE=$1
-PARALLELISM=$2
-export INPUT_BATCH_SIZE=${3:-1000}
-export SENTENCE_WORDS_LIMIT=${4:-120}
+for i in "$@"
+do
+case $i in
+  -in=*|--input=*)
+    INPUT_FILE="${i#*=}"
+    shift
+    ;;
+  -p=*|--parallelism=*)
+    PARALLELISM="${i#*=}"
+    shift
+    ;;
+  -b=*|--batch-size=*)
+    BATCH_SIZE="${i#*=}"
+    shift
+    ;;
+    *)
+    echo "NO MATCH"
+    break
+    ;;
+esac
+done
 
-RUN_SCRIPT=`cd $(dirname $0)/; pwd`/run.sh
+if [ -z "$INPUT_FILE" ]; then
+  echo "Usage: $0 -i=input_file [--parallelism=PARALLELISM] [--batch-size=BATCH_SIZE ] <args for run.sh>"
+  exit
+fi
+
+PARALLELISM=${PARALLELISM:-2}
+BATCH_SIZE=${BATCH_SIZE:-1000}
+
+echo "parallelism = $PARALLELISM"
+echo "batch-size  = $BATCH_SIZE"
+
+RUN_SCRIPT=`cd $(dirname $0)/; pwd`"/run.sh $@"
 echo $RUN_SCRIPT
 mkdir -p $INPUT_FILE.split
 rm -f $INPUT_FILE.split/*
 
 # Split the input file into subfiles
-split -a 10 -l $INPUT_BATCH_SIZE $INPUT_FILE $INPUT_FILE.split/input-
+split -a 10 -l $BATCH_SIZE $INPUT_FILE $INPUT_FILE.split/input-
 
 # Match all files in the split directory
-find $INPUT_FILE.split -name "input-*" 2>/dev/null -print0 | xargs -0 -P $PARALLELISM -L 1 bash -c "${RUN_SCRIPT} -l ${SENTENCE_WORDS_LIMIT}"' -f "$0"'
+find $INPUT_FILE.split -name "input-*" 2>/dev/null -print0 | xargs -0 -P $PARALLELISM -L 1 bash -c "${RUN_SCRIPT}"' -f "$0"'
 
 echo "Output TSV files are in: $INPUT_FILE.split/*.parsed"
 echo "To load them into the databse, run: cat $INPUT_FILE.split/*.parsed | psql YOUR_DB_NAME -c "'"COPY sentences FROM STDIN"'
